Major refactor of Manager, Messages and Parser
Information remaining intact throughout all message-handling stages is
encoded as parameters of nested (case) classes to avoid unnecessarily
long parameter lists of the behavior functions themselves.
Parsers are also restructured in a similar way.

In order to prevent construction of malformed messages, Manager does not
rely on the handlers to properly return the data identifying the task
but creates a layer of adapters associated with particular handlers
which then transparently annotate their responses. (The adapter actors
are memoized.)  Futhermore, the message classes are not defined in
common but under their senders to allow for definition of private
messages.

Printers' lifecycle is watched by Manager and unregistered upon termination.